[MOV] im_livechat: move AbstractThreadWindow outside of -
-public_livechat.js

Task-2212347
